可設置生成武將與已有武將數相關(正比或反比)

+++ GameGlobal/GameGlobal/GlobalVariables.cs
@@ -132,7 +132,7 @@
 
         public static int MaxMilitaryExperience = 3000;
 
-        public static int CreateRandomOfficerChance = 5;
+        public static float CreateRandomOfficerChance = 5;
         public static int GeneratedOfficerFemaleChance = 5;
         public static int ChildrenAvailableAge = 12;
 
@@ -755,7 +755,7 @@
             }
             try
             {
-                CreateRandomOfficerChance = int.Parse(nextSibling.Attributes.GetNamedItem("CreateRandomOfficerChance").Value);
+                CreateRandomOfficerChance = float.Parse(nextSibling.Attributes.GetNamedItem("CreateRandomOfficerChance").Value);
             }
             catch (Exception exception24)
             {
